Добавлены исключения для сервлетов с невалидными id

+++ src/main/java/ru/bikbaev/client_v2/controller/Controller.java
@@ -3,10 +3,12 @@ package ru.bikbaev.client_v2.controller;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import ru.bikbaev.client_v2.dto.DTO;
+import ru.bikbaev.client_v2.exceptions.InvalidIdException;
 import ru.bikbaev.client_v2.mapper.JsonMapper;
 import ru.bikbaev.client_v2.repository.JdbcRepositoryByClient;
 
 import java.io.IOException;
+import java.io.PrintWriter;
 
 public class Controller {
     private static final JsonMapper jsonMapper = new JsonMapper();
@@ -19,15 +21,49 @@ public class Controller {
             Iterable<DTO> dtos = jdbcRepositoryByClient.findAll();
             jsonMapper.printAllJson(dtos, resp);
         } else {
-            int id = Integer.parseInt(idParam);
-            DTO dto = jdbcRepositoryByClient.findById(id);
-            jsonMapper.printJson(dto, resp);
+            try {
+                int id = Integer.parseInt(idParam);
+                if (id < 0) {
+                    throw new InvalidIdException("Invalid id value " + id);
+                } else {
+                    DTO dto = jdbcRepositoryByClient.findById(id);
+                    jsonMapper.printJson(dto, resp);
+                }
+            } catch (NumberFormatException e) {
+                resp.setStatus(500);
+                PrintWriter out = resp.getWriter();
+                out.println("<h1> Invalid id format </h1>");
+
+            } catch (InvalidIdException e) {
+                resp.setStatus(500);
+                PrintWriter out = resp.getWriter();
+                out.println("<h1>" + e.getMessage() + "</h1>");
+            }
+
+
         }
     }
 
-    public void doDelete(HttpServletRequest req,JdbcRepositoryByClient jdbcRepositoryByClient){
+    public void doDelete(HttpServletRequest req,HttpServletResponse resp, JdbcRepositoryByClient jdbcRepositoryByClient) throws IOException {
         String idParam = req.getParameter("id");
-        int id = Integer.parseInt(idParam);
-        jdbcRepositoryByClient.deleteById(id);
+
+
+        try {
+            int id = Integer.parseInt(idParam);
+            if (id < 0) {
+                throw new InvalidIdException("Invalid id value " + id);
+            } else {
+                jdbcRepositoryByClient.deleteById(id);
+            }
+        } catch (NumberFormatException e) {
+            resp.setStatus(500);
+            PrintWriter out = resp.getWriter();
+            out.println("<h1> Invalid id format </h1>");
+
+        } catch (InvalidIdException e) {
+            resp.setStatus(500);
+            PrintWriter out = resp.getWriter();
+            out.println("<h1>" + e.getMessage() + "</h1>");
+        }
     }
 }
